Move function definitions to decorators/
